+++ source/main/preload.js
@@ -6,13 +6,11 @@ import { ipcRenderer as _ipcRenderer, remote as _remote } from 'electron';
 import _electronLog from 'electron-log-daedalus';
 import ElectronStore from 'electron-store';
 import { environment } from './environment';
-import { nodeImplementation } from './config';
+import { nodeImplementation, isFlight } from './config';
 
 const _process = process;
 const _isIncentivizedTestnet = nodeImplementation === 'jormungandr';
-const _electronStore = new ElectronStore({
-  name: _isIncentivizedTestnet ? 'config' : 'config-byron-reboot',
-});
+const _electronStore = new ElectronStore({ name: 'config' });
 
 process.once('loaded', () => {
   Object.assign(global, {
@@ -52,6 +50,7 @@ process.once('loaded', () => {
       platform: os.platform(),
     },
     isIncentivizedTestnet: _isIncentivizedTestnet,
+    isFlight,
   });
   // Expose require for Spectron!
   if (_process.env.NODE_ENV === 'test') {