+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 Changelog
 =========
 
-## 1.0.0
+## 1.0.0-FC1
 
 ### Features
 
+- Integrated network clock api endpoint ([PR 1918](https://github.com/input-output-hk/daedalus/pull/1918))
+- Implements the Flight release information window ([PR 1917](https://github.com/input-output-hk/daedalus/pull/1917))
 - Added new icon for Daedalus Flight ([PR 1909](https://github.com/input-output-hk/daedalus/pull/1909))
+- Integrated network parameters api endpoint ([PR 1915](https://github.com/input-output-hk/daedalus/pull/1915))
 - Implemented flight candidates changes ([PR 1908](https://github.com/input-output-hk/daedalus/pull/1908))
 - Implemented new "Flight candidate" theme ([PR 1914](https://github.com/input-output-hk/daedalus/pull/1914))
 - Implemented non-responding wallets handling ([PR 1901](https://github.com/input-output-hk/daedalus/pull/1901))
@@ -29,6 +32,7 @@ Changelog
 
 ### Chores
 
+- Updated Daedalus name and version for the Daedalus Flight 1.0.0-FC1 release ([PR 1910](https://github.com/input-output-hk/daedalus/pull/1910))
 - Enabled "Wallet UTXO distribution" feature for Byron wallets([PR 1913](https://github.com/input-output-hk/daedalus/pull/1913))
 - Enabled Byron wallet name editing ([PR 1911](https://github.com/input-output-hk/daedalus/pull/1911))
 - Updated UI copy ([PR 1907](https://github.com/input-output-hk/daedalus/pull/1907))